[tagger] Further improvements to remote tagger
We found the remote tagger starts but immediately hangs in some nodes
without anything useful appearing in the logs. We suspect that either
the server is hanging while trying to send a message that the client
never receives, or that the client is receiving it but also getting
stuck.

If that's the case, a newly introduced 1 minute timeout on both ends
should cause the stream to be restared, and hopefully things will get
unstuck on their own. Naturally, we also added some telemetry to track
when that happens.

Additionally, instead of tracking the tagger store state as entities are
added or removed, now we have a separate goroutine running every minute
to take care of that. This avoids not having data for stored entities
when the tagger hangs. It's not all that different in theory, but in
practice we can now detect that in Datadog by querying for agents with
zero stored entities.